Бэта-версия mvc-приложения 1.6.1

+++ index.php
@@ -9,6 +9,7 @@ if(file_exists($path)) {
 
 function startSession($isUserActivity=true) {
     $sessionLifetime = 1800;
+    $url = parse_url($_SERVER['REQUEST_URI']);
     if ( session_id() ) return true;
     ini_set('session.cookie_lifetime', 0);
     ini_set('session.use_strict_mode', 1);
@@ -16,9 +17,8 @@ function startSession($isUserActivity=true) {
     $t = time();
     if ( $sessionLifetime ) {
 
-        if (isset($_SESSION['lastactivity']) && $t-$_SESSION['lastactivity'] >= $sessionLifetime ) {
+        if (isset($_SESSION['lastactivity'])  && $_SESSION['admin'] == 1 && strpos($url['path'], '/admin') !== false  && $t-$_SESSION['lastactivity'] >= $sessionLifetime ) {
                     destroySession();
-                    $url = parse_url($_SERVER['REQUEST_URI']);
                     exit(header( "Refresh:0 url=/admin/login"));
         }
         else {
@@ -31,6 +31,7 @@ function startSession($isUserActivity=true) {
 function destroySession() {
     if ( session_id() ) {
         session_unset();
+        unset($_SESSION['lastactivity']);
         setcookie(session_name(), session_id(), time()-60*60*24);
         session_destroy();
     }
